pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tracecore"
version = "0.1.0"
description = "Production-style AI decision engine built with FastAPI, LangGraph, PostgreSQL, Redis, and PyTorch."
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
  "email-validator>=2.2,<3.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "langchain>=0.3,<0.4",
  "langchain-openai>=0.3,<0.4",
  "langgraph>=0.2,<0.4",
  "pydantic-settings>=2.7,<3.0",
  "PyJWT>=2.10,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "redis>=6.0,<7.0",
  "sqlalchemy>=2.0,<3.0",
  "torch>=2.8,<3.0",
  "uvicorn[standard]>=0.32,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]

